+++ tests/test_files/error_files/py_file_error_invalid_rst.py
@@ -42,9 +42,9 @@ class ExampleClass:
         self.test = "test"
         """Test attr docstring
 
-        .. codeblock:: python
-
-            print("Hello world")
+        = ===
+        1 222
+        == ==
 
         ..